gui: Rename the file gschem_preview.h
