+++ libleptongui/include/preview_widget.h
